Use common attributes for ingested Timestream records

+++ integrations/influxdb_connector/influxdb-timestream-connector/tests/integration_test.rs
@@ -3,6 +3,7 @@ use aws_credential_types::Credentials;
 use aws_sdk_timestreamwrite as timestream_write;
 use aws_types::region::Region;
 use core::time;
+use influxdb_timestream_connector::timestream_utils::retry_with_backoff;
 use influxdb_timestream_connector::{
     records_builder::SchemaType, timestream_utils::TIMESTREAM_API_BASE_WAIT_SECONDS,
 };
@@ -34,17 +35,37 @@ impl CleanupBatch {
 
     async fn cleanup(&mut self, client: &timestream_write::Client) {
         for table_name_to_delete in self.table_names_to_delete.iter() {
+            // Check whether the table exists before trying to delete it.
+            // retry_with_backoff can cause a long wait if the table
+            // doesn't exist
+            let describe_table_result = client
+                .describe_table()
+                .database_name(self.database_name.clone())
+                .table_name(table_name_to_delete)
+                .send()
+                .await;
+            if describe_table_result.is_err() {
+                println!(
+                    "Table to be deleted '{}'.'{}' could not be found",
+                    self.database_name, table_name_to_delete
+                );
+                continue;
+            }
+
             println!(
                 "Deleting table {} in database {}",
                 table_name_to_delete, self.database_name
             );
             thread::sleep(time::Duration::from_secs(TIMESTREAM_API_BASE_WAIT_SECONDS));
-            let result = client
-                .delete_table()
-                .database_name(&self.database_name)
-                .table_name(table_name_to_delete)
-                .send()
-                .await;
+
+            let result = retry_with_backoff(|| {
+                client
+                    .delete_table()
+                    .database_name(&self.database_name)
+                    .table_name(table_name_to_delete)
+                    .send()
+            })
+            .await;
 
             match result {
                 Ok(_) => (),
@@ -117,7 +138,7 @@ async fn test_mtmm_basic() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -158,7 +179,7 @@ async fn test_mtmm_create_database() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -181,12 +202,25 @@ async fn test_mtmm_create_database() -> Result<(), Error> {
         test_create_database_name.to_string(),
         vec![lp_measurement_name],
     );
+    // Deletes tables
     cleanup_batch.cleanup(&client).await;
-    let database_delete_response = client
-        .delete_database()
+
+    // Manually delete the created database
+    let describe_database_result = client
+        .describe_database()
         .database_name(test_create_database_name)
         .send()
         .await;
+    assert!(describe_database_result.is_ok());
+
+    let database_delete_response = retry_with_backoff(|| {
+        client
+            .delete_database()
+            .database_name(test_create_database_name)
+            .send()
+    })
+    .await;
+
     if database_delete_response.is_err() {
         println!(
             "The database {} failed to delete",
@@ -209,7 +243,7 @@ async fn test_mtmm_unusual_query_parameters() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -246,7 +280,7 @@ async fn test_mtmm_no_query_parameters() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -281,7 +315,7 @@ async fn test_mtmm_multiple_timestamps() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {} {}\n",
@@ -314,7 +348,7 @@ async fn test_mtmm_many_tags_many_fields() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let mut point = format!("{},", lp_measurement_name);
 
@@ -363,7 +397,7 @@ async fn test_mtmm_float() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={} {}\n",
@@ -400,7 +434,7 @@ async fn test_mtmm_string() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1=\"{}\" {}\n",
@@ -437,7 +471,7 @@ async fn test_mtmm_bool() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={} {}\n",
@@ -477,7 +511,7 @@ async fn test_mtmm_max_tag_length() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},{}={} field1={}i {}\n",
@@ -518,7 +552,7 @@ async fn test_mtmm_beyond_max_tag_length() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},{}={} field1={}i {}\n",
@@ -553,7 +587,7 @@ async fn test_mtmm_max_field_length() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},field1={} {}=\"{}\" {}\n",
@@ -593,7 +627,7 @@ async fn test_mtmm_beyond_max_field_length() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},field1={} {}=\"{}\" {}\n",
@@ -631,7 +665,7 @@ async fn test_mtmm_max_unique_field_keys() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let mut lp_batch = String::new();
     for i in 0..1024 {
@@ -675,7 +709,7 @@ async fn test_mtmm_beyond_max_unique_field_keys() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let mut lp_batch = String::new();
     for i in 0..1025 {
@@ -718,7 +752,7 @@ async fn test_mtmm_max_unique_tag_keys() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let mut lp_batch = String::new();
     for i in 0..128 {
@@ -762,7 +796,7 @@ async fn test_mtmm_beyond_max_unique_tag_keys() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let mut lp_batch = String::new();
     for i in 0..129 {
@@ -876,7 +910,7 @@ async fn test_mtmm_nanosecond_precision() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -915,7 +949,7 @@ async fn test_mtmm_microsecond_precision() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -952,7 +986,7 @@ async fn test_mtmm_second_precision() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -989,7 +1023,7 @@ async fn test_mtmm_no_precision() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1054,7 +1088,7 @@ pub async fn test_mtmm_small_timestamp() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1092,7 +1126,7 @@ async fn test_mtmm_5_measurements() -> Result<(), Error> {
     let client = Arc::new(client);
 
     let mut table_names_to_delete = Vec::<String>::new();
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let mut lp_batch = String::new();
     for i in 0..5 {
@@ -1138,7 +1172,7 @@ async fn test_mtmm_100_measurements() -> Result<(), Error> {
 
     let mut table_names_to_delete = Vec::<String>::new();
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
     let mut lp_batch = String::new();
     for i in 0..100 {
         let lp_measurement_name = format!("{lp_measurement_name}{i}").to_string();
@@ -1183,7 +1217,7 @@ async fn test_mtmm_5000_batch() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
     let mut lp_batch = String::new();
 
     for _ in 0..5000 {
@@ -1234,7 +1268,7 @@ async fn test_mtmm_no_credentials() {
     let client = Arc::new(client);
     tokio::task::spawn(reload.reload_task());
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1279,7 +1313,7 @@ async fn test_mtmm_incorrect_credentials() {
     let client = Arc::new(client);
     tokio::task::spawn(reload.reload_task());
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1312,7 +1346,7 @@ async fn test_mtmm_custom_dimension_partition_key_optional_enforcement() -> Resu
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1354,7 +1388,7 @@ async fn test_mtmm_custom_dimension_partition_key_required_enforcement_accepted(
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1396,7 +1430,7 @@ async fn test_mtmm_custom_dimension_partition_key_required_enforcement_rejected(
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1436,7 +1470,7 @@ async fn test_mtmm_custom_dimension_partition_key_no_dimension() -> Result<(), E
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1476,7 +1510,7 @@ async fn test_mtmm_custom_dimension_partition_key_no_enforcement() -> Result<(),
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1516,7 +1550,7 @@ async fn test_mtmm_custom_measure_partition_key() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1557,7 +1591,7 @@ async fn test_mtmm_custom_measure_partition_key_with_dimension() -> Result<(), E
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1599,7 +1633,7 @@ async fn test_mtmm_custom_measure_partition_key_with_enforcement() -> Result<(),
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1636,7 +1670,7 @@ async fn test_stmm_basic() -> Result<(), Error> {
         .expect("Failed to get client");
     let client = Arc::new(client);
 
-    let lp_measurement_name = String::from("readings");
+    let lp_measurement_name = random_string(11);
 
     let point = format!(
         "{},tag1={} field1={}i {}\n",
@@ -1675,7 +1709,7 @@ async fn test_stmm_varying_metrics() -> Result<(), Error> {
 
     let mut table_names_to_delete = Vec::<String>::new();
 
-    let lp_readings_measurement_name = String::from("readings");
+    let lp_readings_measurement_name = random_string(11);
     let mut lp_batch = String::new();
     for i in 0..10 {
         let lp_readings_measurement_name = format!("{lp_readings_measurement_name}{i}").to_string();
